add credential
