+++ sdk/identity/identity/src/credentials/azurePipelinesServiceConnectionCredentialOptions.ts
@@ -1,0 +1,14 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { AuthorityValidationOptions } from "./authorityValidationOptions";
+import { CredentialPersistenceOptions } from "./credentialPersistenceOptions";
+import { MultiTenantTokenCredentialOptions } from "./multiTenantTokenCredentialOptions";
+
+/**
+ * Optional parameters for the {@link AzurePipelinesServiceConnectionCredential} class.
+ */
+export interface AzurePipelinesServiceConnectionCredentialOptions
+  extends MultiTenantTokenCredentialOptions,
+    CredentialPersistenceOptions,
+    AuthorityValidationOptions {}
