+++ sdk/identity/identity/src/credentials/workloadIdentityCredential.ts
@@ -8,11 +8,6 @@ import { readFile } from "fs/promises";
 import { CredentialUnavailableError } from "../errors";
 import { credentialLogger, processEnvVars } from "../util/logging";
 import { checkTenantId } from "../util/tenantIdUtils";
-import {
-  createDefaultHttpClient,
-  createHttpHeaders,
-  createPipelineRequest,
-} from "@azure/core-rest-pipeline";
 
 const credentialName = "WorkloadIdentityCredential";
 /**
@@ -47,7 +42,6 @@ export class WorkloadIdentityCredential implements TokenCredential {
   private azureFederatedTokenFileContent: string | undefined = undefined;
   private cacheDate: number | undefined = undefined;
   private federatedTokenFilePath: string | undefined;
-  private serviceConnectionId: string | undefined;
 
   /**
    * WorkloadIdentityCredential supports Microsoft Entra Workload ID on Kubernetes.
@@ -62,66 +56,21 @@ export class WorkloadIdentityCredential implements TokenCredential {
     const workloadIdentityCredentialOptions = options ?? {};
     const tenantId = workloadIdentityCredentialOptions.tenantId || process.env.AZURE_TENANT_ID;
     const clientId = workloadIdentityCredentialOptions.clientId || process.env.AZURE_CLIENT_ID;
-
     this.federatedTokenFilePath =
-      workloadIdentityCredentialOptions?.tokenFilePath || process.env.AZURE_FEDERATED_TOKEN_FILE;
-    this.serviceConnectionId = workloadIdentityCredentialOptions.serviceConnectionId;
-    if (!tenantId || !clientId) {
-      throw new CredentialUnavailableError(
-        `${credentialName}: is Unavailable. clientId and tenantId are required parameters/ environment varables - AZURE_TENANT_ID, AZURE_CLIENT_ID`
-      );
-    }
+      workloadIdentityCredentialOptions.tokenFilePath || process.env.AZURE_FEDERATED_TOKEN_FILE;
     if (tenantId) {
       checkTenantId(logger, tenantId);
     }
-    if (this.federatedTokenFilePath && this.serviceConnectionId) {
-      const errorMessage = `${credentialName}: ambiguous. serviceConnectionId and tokenFilePath cannot be provided at the same time. These are used for supporting Workload Identity for different environments. 
-      The tokenFilePath is used for Kubernetes while serviceConnectionId is used for Azure Pipelines.
-      See the troubleshooting guide for more information: https://aka.ms/azsdk/js/identity/workloadidentitycredential/troubleshoot`;
-      logger.error(errorMessage);
-      throw new CredentialUnavailableError(errorMessage);
-    }
-
     if (clientId && tenantId && this.federatedTokenFilePath) {
       logger.info(
-        `Invoking ClientAssertionCredential with tenant ID: ${tenantId}, clientId: ${workloadIdentityCredentialOptions.clientId} and federated token path: [REDACTED]`
+        `Invoking ClientAssertionCredential with tenant ID: ${tenantId}, clientId: ${workloadIdentityCredentialOptions.clientId} and federated token path: [REDACTED]`,
       );
       this.client = new ClientAssertionCredential(
         tenantId,
         clientId,
         this.readFileContents.bind(this),
-        options
+        options,
       );
-    } else {
-      if (clientId && tenantId && this.serviceConnectionId) {
-        //Ensure all system env vars are there to form the request uri for OIDC token
-        this.ensurePipelinesSystemVars();
-        const oidcRequestUrl = `${process.env.SYSTEM_TEAMFOUNDATIONCOLLECTIONURI}${process.env.SYSTEM_TEAMPROJECTID}/_apis/distributedtask/hubs/build/plans/${process.env.SYSTEM_PLANID}/jobs/${process.env.SYSTEM_JOBID}/oidctoken?api-version=7.1-preview.1&serviceConnectionId=${this.serviceConnectionId}`;
-        const systemAccessToken = `${process.env.SYSTEM_ACCESSTOKEN}`;
-        logger.info(
-          `Invoking ClientAssertionCredential with tenant ID: ${tenantId}, clientId: ${workloadIdentityCredentialOptions.clientId} and service connection id: ${workloadIdentityCredentialOptions.serviceConnectionId}`
-        );
-        this.client = new ClientAssertionCredential(
-          tenantId,
-          clientId,
-          this.requestOidcToken.bind(this, oidcRequestUrl, systemAccessToken),
-          options
-        );
-      } else {
-        const errorMessage = `${credentialName}: is unavailable. tenantId, clientId, and either tokenFilePath or serviceConnectionId are required parameters. 
-      To enable Workload Identity Federation please provide following environment variables based on the environment.
-      For Azure Pipelines env, in WorkloadIdentityCredential, these are required as inputs / env variables - 
-      "AZURE_TENANT_ID",
-      "AZURE_CLIENT_ID",
-      serviceConnectionId.
-      For Kubernetes env, in DefaultAzureCredential and ManagedIdentityCredential, these can be provided as environment variables - 
-      "AZURE_TENANT_ID",
-      "AZURE_CLIENT_ID",
-      "AZURE_FEDERATED_TOKEN_FILE". 
-      See the troubleshooting guide for more information: https://aka.ms/azsdk/js/identity/workloadidentitycredential/troubleshoot`;
-        logger.error(errorMessage);
-        throw new CredentialUnavailableError(errorMessage);
-      }
     }
   }
 
@@ -135,21 +84,15 @@ export class WorkloadIdentityCredential implements TokenCredential {
    */
   public async getToken(
     scopes: string | string[],
-    options?: GetTokenOptions
+    options?: GetTokenOptions,
   ): Promise<AccessToken | null> {
     if (!this.client) {
-      const errorMessage = `${credentialName}: is unavailable. tenantId, clientId, and either tokenFilePath or serviceConnectionId are required parameters. 
-      To enable Workload Identity Federation please provide following environment variables based on the environment.
-      For Azure Pipelines env, in WorkloadIdentityCredential, these are required as inputs / env variables - 
+      const errorMessage = `${credentialName}: is unavailable. tenantId, clientId, and federatedTokenFilePath are required parameters. 
+      In DefaultAzureCredential and ManagedIdentityCredential, these can be provided as environment variables - 
       "AZURE_TENANT_ID",
       "AZURE_CLIENT_ID",
-      serviceConnectionId.
-      For Kubernetes env, in DefaultAzureCredential and ManagedIdentityCredential, these can be provided as environment variables - 
-      "AZURE_TENANT_ID",
-      "AZURE_CLIENT_ID",
-      "AZURE_FEDERATED_TOKEN_FILE". 
-      See the troubleshooting guide for more information: https://aka.ms/azsdk/js/identity/workloadidentitycredential/troubleshoot`;
-      logger.error(errorMessage);
+      "AZURE_FEDERATED_TOKEN_FILE". See the troubleshooting guide for more information: https://aka.ms/azsdk/js/identity/workloadidentitycredential/troubleshoot  `;
+      logger.info(errorMessage);
       throw new CredentialUnavailableError(errorMessage);
     }
     logger.info("Invoking getToken() of Client Assertion Credential");
@@ -163,7 +106,7 @@ export class WorkloadIdentityCredential implements TokenCredential {
     }
     if (!this.federatedTokenFilePath) {
       throw new CredentialUnavailableError(
-        `${credentialName}: is unavailable. Invalid file path provided ${this.federatedTokenFilePath}.`
+        `${credentialName}: is unavailable. Invalid file path provided ${this.federatedTokenFilePath}.`,
       );
     }
     if (!this.azureFederatedTokenFileContent) {
@@ -171,7 +114,7 @@ export class WorkloadIdentityCredential implements TokenCredential {
       const value = file.trim();
       if (!value) {
         throw new CredentialUnavailableError(
-          `${credentialName}: is unavailable. No content on the file ${this.federatedTokenFilePath}.`
+          `${credentialName}: is unavailable. No content on the file ${this.federatedTokenFilePath}.`,
         );
       } else {
         this.azureFederatedTokenFileContent = value;
@@ -180,73 +123,4 @@ export class WorkloadIdentityCredential implements TokenCredential {
     }
     return this.azureFederatedTokenFileContent;
   }
-
-  private async requestOidcToken(
-    oidcRequestUrl: string,
-    systemAccessToken: string
-  ): Promise<string> {
-    logger.info("Requesting OIDC token from Azure Pipelines...");
-    logger.info(oidcRequestUrl);
-
-    const httpClient = createDefaultHttpClient();
-
-    const request = createPipelineRequest({
-      url: oidcRequestUrl,
-      method: "POST",
-      headers: createHttpHeaders({
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${systemAccessToken}`,
-      }),
-    });
-
-    const response = await httpClient.sendRequest(request);
-    const text = response.bodyAsText;
-    if (!text) {
-      throw new CredentialUnavailableError(
-        `${credentialName}: is unavailable. Received null token from OIDC request. Response status = ${response.status}`
-      );
-    }
-    const result = JSON.parse(text);
-    if (result?.oidcToken) {
-      return result.oidcToken;
-    } else {
-      throw new CredentialUnavailableError(
-        `${credentialName}: is unavailable. oidcToken field not detected in the response. Response = ${JSON.stringify(
-          result
-        )}`
-      );
-    }
-  }
-
-  private ensurePipelinesSystemVars(): void {
-    if (
-      process.env.SYSTEM_TEAMFOUNDATIONCOLLECTIONURI &&
-      process.env.SYSTEM_TEAMPROJECTID &&
-      process.env.SYSTEM_PLANID &&
-      process.env.SYSTEM_JOBID &&
-      process.env.SYSTEM_ACCESSTOKEN
-    ) {
-      return;
-    }
-    const missingEnvVars = [];
-    if (!process.env.SYSTEM_TEAMFOUNDATIONCOLLECTIONURI)
-      missingEnvVars.push("SYSTEM_TEAMFOUNDATIONCOLLECTIONURI");
-    if (!process.env.SYSTEM_TEAMFOUNDATIONCOLLECTIONURI)
-      missingEnvVars.push("SYSTEM_TEAMFOUNDATIONCOLLECTIONURI");
-    if (!process.env.SYSTEM_TEAMPROJECTID) missingEnvVars.push("SYSTEM_TEAMPROJECTID");
-    if (!process.env.SYSTEM_PLANID) missingEnvVars.push("SYSTEM_PLANID");
-    if (!process.env.SYSTEM_JOBID) missingEnvVars.push("SYSTEM_JOBID");
-    if (!process.env.SYSTEM_ACCESSTOKEN) missingEnvVars.push("SYSTEM_ACCESSTOKEN");
-    if (missingEnvVars.length > 0)
-      throw new CredentialUnavailableError(
-        `${credentialName}: is unavailable. Missing system variable(s) - ${missingEnvVars.join(
-          ", "
-        )}`
-      );
-  }
 }
-
-// should we skip this check? https://microsoft.visualstudio.com/Edge/_git/edgeinternal.es?path=/UtilityLibraries/Microsoft.Edge.ES.Azure.Identity/AzureDevOpsFederatedTokenCredential.cs&version=GBmaster&line=124&lineEnd=125&lineStartColumn=1&lineEndColumn=1&lineStyle=plain&_a=contents
-// to not to tightly couple to Devops implememntation in case they make changes in host support
-
-// check if system variables can be overridden
